Only store failed tests in the ExUnit manifest
The only tests we care about are the failed ones, so we can
ignore passed tests and not bother storing the `last_run_status`.
Once we have done that, the only associated piece of data we
care about is the file, so we can replace the `entry` record with
just the file.